Removed UseEffect From searchBar and SidePanel files

+++ backend/src/main/java/com/example/nr/quickports/services/ChargeDeviceServiceInterface.java
@@ -9,7 +9,7 @@ public interface ChargeDeviceServiceInterface {
 
     ChargeDevice deleteChargeDeviceById(Long id);
 
-    List<ChargeDevice> findChargeDevicesByDistance(LatLng coordinateLocation , Long Distance);
+    List<ChargeDevice> findChargeDevicesByDistance(LatLng coordinateLocation , double Distance);
